docs(card): update card docs and structure
