+++ .storybook/main.js
@@ -16,7 +16,8 @@ module.exports = {
     '../projects/canopy/src/lib/button/docs/guide.stories.mdx',
     '../projects/canopy/src/lib/button/docs/button.stories.ts',
     '../projects/canopy/src/lib/button/docs/button-group.stories.ts',
-    '../projects/canopy/src/lib/card/card.stories.ts',
+    '../projects/canopy/src/lib/card/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/card/docs/card.stories.ts',
     '../projects/canopy/src/lib/carousel/carousel.stories.ts',
     '../projects/canopy/src/lib/data-point/data-point.stories.ts',
     '../projects/canopy/src/lib/details/docs/guide.stories.mdx',
