+++ projects/canopy/src/lib/accordion/docs/guide.stories.mdx
@@ -113,7 +113,7 @@ This component is based on best practice and has not had specific usability test
 
 Consider using an accordion instead of the **details component** if there are multiple related sections of content. **Tabs** would be an alternative way to display if there's a large amount of content.
 
-- [Card](./?path=/docs/components-card--default-card)
+- [Card](./?path=/docs/components-card-guide--page)
 - [Details](./?path=/docs/components-details-guide--page)
 - [Tabs](./?path=/docs/components-tabs-tabbed-content--standard-tabs)
